+++ core/src/test/java/com/graphhopper/routing/util/Bike2WeightTagParserTest.java
@@ -42,8 +42,8 @@ public class Bike2WeightTagParserTest extends BikeTagParserTest {
     }
 
     @Override
-    protected BikeCommonTagParser createBikeTagParser(EncodedValueLookup lookup) {
-        Bike2WeightTagParser parser = new Bike2WeightTagParser(lookup, new PMap("block_fords=true"));
+    protected BikeCommonTagParser createBikeTagParser(EncodedValueLookup lookup, PMap pMap) {
+        Bike2WeightTagParser parser = new Bike2WeightTagParser(lookup, pMap);
         parser.init(new DateRangeParser());
         return parser;
     }
@@ -119,7 +119,7 @@ public class Bike2WeightTagParserTest extends BikeTagParserTest {
         way.setTag("route", "ferry");
         way.setTag("edge_distance", 500.0);
 
-        assertNotEquals(EncodingManager.Access.CAN_SKIP, parser.getAccess(way));
+        assertNotEquals(WayAccess.CAN_SKIP, parser.getAccess(way));
         IntsRef edgeFlags = encodingManager.createEdgeFlags();
         edgeFlags = osmParsers.handleWayTags(edgeFlags, way, encodingManager.createRelationFlags());
         graph.edge(0, 1).setDistance(247).setFlags(edgeFlags);