+++ core/src/test/java/com/graphhopper/routing/util/BikeTagParserTest.java
@@ -47,8 +47,8 @@ public class BikeTagParserTest extends AbstractBikeTagParserTester {
     }
 
     @Override
-    protected BikeCommonTagParser createBikeTagParser(EncodedValueLookup lookup) {
-        BikeTagParser parser = new BikeTagParser(lookup, new PMap("block_fords=true"));
+    protected BikeCommonTagParser createBikeTagParser(EncodedValueLookup lookup, PMap pMap) {
+        BikeTagParser parser = new BikeTagParser(lookup, pMap);
         parser.init(new DateRangeParser());
         return parser;
     }
@@ -597,7 +597,7 @@ public class BikeTagParserTest extends AbstractBikeTagParserTester {
         way = new ReaderWay(1);
         way.setTag("highway", "residential");
         way.setTag("maxspeed", "15");
-        // todonow!! speed is larger than maxspeed
+        // todo: speed is larger than maxspeed tag due to rounding and storable max speed is 30
         assertPriorityAndSpeed(PREFER.getValue(), 16, way);
     }